# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for ordering food.",
        "title": "Food Ordering API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/orders": {
            "post": {
                "description": "สร้างออเดอร์ใหม่พร้อมรายการอาหารและโปรโมชั่น (ถ้ามี)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order_ใหม่"
                ],
                "summary": "สร้างออเดอร์ใหม่",
                "parameters": [
                    {
                        "description": "ข้อมูลออเดอร์",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_handlers.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                }
            }
        },
...
        "/api/orders/items/serve/{id}": {
            "post": {
                "description": "ยืนยันการเสิร์ฟอาหารสำหรับรายการอาหารในออเดอร์",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order_ใหม่"
                ],
                "summary": "ยืนยันการเสิร์ฟอาหาร",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "OrderItem ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OrderItem"
                        }
                    }
                }
            }
        },
        "/api/orders/status/{id}": {
            "put": {
                "description": "อัพเดทสถานะของออเดอร์ (pending, preparing, ready, served, cancelled)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order_ใหม่"
                ],
                "summary": "อัพเดทสถานะออเดอร์",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "สถานะใหม่",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_handlers.updateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                }
            }
        },
        "/api/orders/table/{table_id}": {
            "get": {
                "description": "ดึงรายการออเดอร์ทั้งหมดของโต๊ะที่ระบุ",
                "produces": [
                    "application/json"
                ],
                "summary": "ดึงรายการออเดอร์ตามโต๊ะ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Table ID",
                        "name": "table_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Order"
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "description": "ดึงรายละเอียดของออเดอร์ตาม ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order_ใหม่"
                ],
                "summary": "ดึงรายละเอียดออเดอร์",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                }
            }
        },
        "/api/orders/{id}/cancel": {
            "post": {
                "description": "ยกเลิกออเดอร์และรายการอาหารทั้งหมดในออเดอร์",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ยกเลิกออเดอร์",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api_handlers.CreateOrderRequest": {
            "type": "object",
            "required": [
                "items",
                "table_id",
                "uuid"
            ],
            "properties": {
                "items": {
                    "description": "รายการอาหารที่สั่ง",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api_handlers.orderItemRequest"
                    }
                },
                "table_id": {
                    "description": "ID ของโต๊ะ",
                    "type": "integer"
                },
                "use_promo": {
                    "description": "โปรโมชั่นที่ใช้ (ถ้ามี)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api_handlers.UsePromoRequest"
                    }
                },
                "uuid": {
                    "description": "UUID ของ QR Code",
                    "type": "string"
                }
            }
        },
...
        "api_handlers.UsePromoRequest": {
            "type": "object",
            "required": [
                "menu_item_ids",
                "promotion_id"
            ],
            "properties": {
                "menu_item_ids": {
                    "description": "รายการอาหารที่ใช้โปรโมชั่น",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "promotion_id": {
                    "type": "integer"
                }
            }
        },
...
        "api_handlers.orderItemRequest": {
            "type": "object",
            "required": [
                "menu_item_id",
                "quantity"
            ],
            "properties": {
                "menu_item_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api_handlers.OrderItemOptionRequest"
                    }
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "api_handlers.updateStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "preparing",
                        "ready",
                        "served",
                        "cancelled"
                    ]
                }
            }
        },
...
        "models.Order": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderItem"
                    }
                },
                "receipt": {
                    "$ref": "#/definitions/models.Receipt"
                },
                "receiptID": {
                    "type": "integer"
                },
                "status": {
                    "description": "\"completed\", \"uncompleted\"",
                    "type": "string"
                },
                "tableID": {
                    "type": "integer"
                },
                "total": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "models.OrderItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menuItem": {
                    "$ref": "#/definitions/models.MenuItem"
                },
                "menuItemID": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "options": {
                    "description": "เพิ่ม relation กับ options",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderItemOption"
                    }
                },
                "order": {
                    "$ref": "#/definitions/models.Order"
                },
                "orderID": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "servedAt": {
                    "description": "เวลาที่เสิร์ฟอาหาร",
                    "type": "string"
                },
                "status": {
                    "description": "pending, served, cancelled",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.OrderItemOption": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menuOption": {
                    "$ref": "#/definitions/models.MenuOption"
                },
                "menuOptionID": {
                    "description": "Foreign key to MenuOption",
                    "type": "integer"
                },
                "orderItem": {
                    "$ref": "#/definitions/models.OrderItem"
                },
                "orderItemID": {
                    "description": "Foreign key to OrderItem",
                    "type": "integer"
                },
                "price": {
                    "description": "ราคา ณ เวลาที่สั่ง",
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "description": "ค่าที่เลือก เช่น \"เผ็ดมาก\", \"เพิ่มไข่ดาว\"",
                    "type": "string"
                }
            }
        },
...
        "models.Receipt": {
            "type": "object",
            "properties": {
                "chargeTotal": {
                    "type": "number"
                },
                "charges": {
                    "description": "เปลี่ยนจาก OrderAdditionalCharge",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ReceiptCharge"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "discountTotal": {
                    "type": "number"
                },
                "discounts": {
                    "description": "เปลี่ยนจาก OrderDiscount เพราะมันไม่ตอบโจทย์ T T",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ReceiptDiscount"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Order"
                    }
                },
                "paymentMethod": {
                    "type": "string"
                },
                "serviceCharge": {
                    "type": "number"
                },
                "staff": {
                    "$ref": "#/definitions/models.Users"
                },
                "staffID": {
                    "type": "integer"
                },
                "subTotal": {
                    "description": "ยอดรวมทุก order",
                    "type": "number"
                },
                "tableID": {
                    "type": "integer"
                },
                "total": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}